feat: Add concat array support addressing all review comments
- Move concat_arrays from datafusion-common to datafusion-functions/src/utils.rs
- Remove unnecessary pub declarations and thin wrapper functions
- Restore find_or_first logic and missing align_array_dimensions test
- Add documentation for string type precedence and PostgreSQL compatibility
- Use ColumnarValue::values_to_arrays() instead of manual conversion
- Simplify return_type/invoke logic to only check first argument after coercion
- Fix empty argument handling to require at least one argument

Implements array concatenation: concat([1,2], [3,4]) → [1,2,3,4]
Supports various data types and multi-dimensional arrays
