+++ datafusion/common/src/utils/mod.rs
@@ -17,7 +17,6 @@
 
 //! This module provides the bisect function, which implements binary search.
 
-pub mod array_utils;
 pub mod expr;
 pub mod memory;
 pub mod proxy;
