+++ datafusion/functions/src/utils.rs
@@ -15,13 +15,20 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use arrow::array::{Array, ArrayRef, ArrowPrimitiveType, AsArray, PrimitiveArray};
+use arrow::array::{
+    Array, ArrayData, ArrayRef, ArrowPrimitiveType, AsArray, GenericListArray,
+    NullBufferBuilder, OffsetSizeTrait, PrimitiveArray,
+};
+use arrow::buffer::OffsetBuffer;
 use arrow::compute::try_binary;
-use arrow::datatypes::{DataType, DecimalType};
+use arrow::datatypes::{DataType, DecimalType, Field};
 use arrow::error::ArrowError;
-use datafusion_common::{not_impl_err, DataFusionError, Result, ScalarValue};
+use datafusion_common::cast::as_generic_list_array;
+use datafusion_common::utils::list_ndims;
+use datafusion_common::{exec_err, not_impl_err, DataFusionError, Result, ScalarValue};
 use datafusion_expr::function::Hint;
 use datafusion_expr::ColumnarValue;
+use itertools::Itertools;
 use std::sync::Arc;
 
 /// Creates a function to identify the optimal return type of a string function given
@@ -376,5 +383,203 @@ pub mod test {
                 Err(_) => assert!(expected.is_none()),
             }
         }
+    }
+}
+
+/// Concatenates arrays
+pub fn concat_arrays(args: &[ArrayRef]) -> Result<ArrayRef> {
+    if args.is_empty() {
+        return exec_err!("concat_arrays expects at least one argument");
+    }
+
+    let mut all_null = true;
+    let mut large_list = false;
+    for arg in args {
+        match arg.data_type() {
+            DataType::Null => continue,
+            DataType::LargeList(_) => large_list = true,
+            _ => (),
+        }
+        if arg.null_count() < arg.len() {
+            all_null = false;
+        }
+    }
+
+    if all_null {
+        // Return a null array with the same type as the first non-null-type argument
+        // Safe because args is non-empty
+        let return_type = args
+            .iter()
+            .map(|arg| arg.data_type())
+            .find_or_first(|d| !d.is_null())
+            .unwrap();
+
+        return Ok(arrow::array::make_array(ArrayData::new_null(
+            return_type,
+            args[0].len(),
+        )));
+    }
+
+    if large_list {
+        concat_arrays_internal::<i64>(args)
+    } else {
+        concat_arrays_internal::<i32>(args)
+    }
+}
+
+fn concat_arrays_internal<O: OffsetSizeTrait>(args: &[ArrayRef]) -> Result<ArrayRef> {
+    let args = align_array_dimensions::<O>(args.to_vec())?;
+
+    let list_arrays = args
+        .iter()
+        .map(|arg| as_generic_list_array::<O>(arg))
+        .collect::<Result<Vec<_>>>()?;
+
+    // Assume number of rows is the same for all arrays
+    let row_count = list_arrays[0].len();
+
+    let mut array_lengths = vec![];
+    let mut arrays = vec![];
+    let mut valid = NullBufferBuilder::new(row_count);
+    for i in 0..row_count {
+        let nulls = list_arrays
+            .iter()
+            .map(|arr| arr.is_null(i))
+            .collect::<Vec<_>>();
+
+        // If all the arrays are null, the concatenated array is null
+        let is_null = nulls.iter().all(|&x| x);
+        if is_null {
+            array_lengths.push(0);
+            valid.append_null();
+        } else {
+            // Get all the arrays on i-th row
+            let values = list_arrays
+                .iter()
+                .map(|arr| arr.value(i))
+                .collect::<Vec<_>>();
+
+            let elements = values
+                .iter()
+                .map(|a| a.as_ref())
+                .collect::<Vec<&dyn Array>>();
+
+            // Concatenated array on i-th row
+            let concatenated_array = arrow::compute::concat(elements.as_slice())?;
+            array_lengths.push(concatenated_array.len());
+            arrays.push(concatenated_array);
+            valid.append_non_null();
+        }
+    }
+    // Assume all arrays have the same data type
+    let data_type = list_arrays[0].value_type();
+
+    let elements = arrays
+        .iter()
+        .map(|a| a.as_ref())
+        .collect::<Vec<&dyn Array>>();
+
+    let list_arr = GenericListArray::<O>::new(
+        Arc::new(Field::new_list_field(data_type, true)),
+        OffsetBuffer::from_lengths(array_lengths),
+        Arc::new(arrow::compute::concat(elements.as_slice())?),
+        valid.finish(),
+    );
+
+    Ok(Arc::new(list_arr))
+}
+
+/// Aligns array dimensions
+fn align_array_dimensions<O: OffsetSizeTrait>(
+    args: Vec<ArrayRef>,
+) -> Result<Vec<ArrayRef>> {
+    let args_ndim = args
+        .iter()
+        .map(|arg| list_ndims(arg.data_type()))
+        .collect::<Vec<_>>();
+    let max_ndim = args_ndim.iter().max().unwrap_or(&0);
+
+    // Align the dimensions of the arrays
+    let aligned_args: Result<Vec<ArrayRef>> = args
+        .into_iter()
+        .zip(args_ndim.iter())
+        .map(|(array, ndim)| {
+            if ndim < max_ndim {
+                let mut aligned_array = Arc::clone(&array);
+                for _ in 0..(max_ndim - ndim) {
+                    let data_type = aligned_array.data_type().to_owned();
+                    let array_lengths = vec![1; aligned_array.len()];
+                    let offsets = OffsetBuffer::<O>::from_lengths(array_lengths);
+
+                    let field = Arc::new(Field::new("item", data_type, true));
+                    let aligned_array_inner =
+                        GenericListArray::<O>::new(field, offsets, aligned_array, None);
+                    aligned_array = Arc::new(aligned_array_inner);
+                }
+                Ok(aligned_array)
+            } else {
+                Ok(array)
+            }
+        })
+        .collect();
+
+    aligned_args
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use arrow::array::ListArray;
+    use arrow::datatypes::Int64Type;
+    use datafusion_common::cast::as_list_array;
+    use datafusion_common::utils::{list_ndims, SingleRowListArrayBuilder};
+
+    /// Test for align_array_dimensions function
+    #[test]
+    fn test_align_array_dimensions() {
+        let array1d_1: ArrayRef =
+            Arc::new(ListArray::from_iter_primitive::<Int64Type, _, _>(vec![
+                Some(vec![Some(1), Some(2), Some(3)]),
+                Some(vec![Some(4), Some(5)]),
+            ]));
+        let array1d_2: ArrayRef =
+            Arc::new(ListArray::from_iter_primitive::<Int64Type, _, _>(vec![
+                Some(vec![Some(6), Some(7), Some(8)]),
+            ]));
+
+        let array2d_1: ArrayRef = Arc::new(
+            SingleRowListArrayBuilder::new(Arc::clone(&array1d_1)).build_list_array(),
+        );
+        let array2d_2 = Arc::new(
+            SingleRowListArrayBuilder::new(Arc::clone(&array1d_2)).build_list_array(),
+        );
+
+        let res = align_array_dimensions::<i32>(vec![
+            array1d_1.to_owned(),
+            array2d_2.to_owned(),
+        ])
+        .unwrap();
+
+        let expected = as_list_array(&array2d_1).unwrap();
+        let expected_dim = list_ndims(array2d_1.data_type());
+        assert_ne!(as_list_array(&res[0]).unwrap(), expected);
+        assert_eq!(
+            list_ndims(res[0].data_type()),
+            expected_dim
+        );
+
+        let array3d_1: ArrayRef =
+            Arc::new(SingleRowListArrayBuilder::new(array2d_1).build_list_array());
+        let array3d_2: ArrayRef =
+            Arc::new(SingleRowListArrayBuilder::new(array2d_2).build_list_array());
+        let res = align_array_dimensions::<i32>(vec![array1d_1, array3d_2]).unwrap();
+
+        let expected = as_list_array(&array3d_1).unwrap();
+        let expected_dim = list_ndims(array3d_1.data_type());
+        assert_ne!(as_list_array(&res[0]).unwrap(), expected);
+        assert_eq!(
+            list_ndims(res[0].data_type()),
+            expected_dim
+        );
     }
 }
